fix : orgnazing app folder & fix naming
