+++ vet_nest/src/diagnosis/diagnosis.module.ts
@@ -1,7 +1,6 @@
 import { Module } from '@nestjs/common';
 import { DisgnosisController } from './diagnosis.controller';
 
-import { AppService } from '@/app.service';
 import { DatabaseModule } from '@/database/database.module';
 import { DiagnosisService } from './diagnosis.service';
