+++ vet_nest/src/exceptions/exceptions.module.ts
@@ -1,4 +1,4 @@
 import { Module } from '@nestjs/common';
 
 @Module({})
-export class ExceptionModule {}
+export class ExceptionsModule {}
